irc: use NAMES instead of WHO
Refactor Capabilities tracking and use NAMES instead of WHO. WHO is rate
limited and comes with more information. Begin tracking of ACK'd caps so
that we can (in future commits) conditionally use WHOX and
account-notify features to request AWAY information only when a server
implements WHOX (IE soju, which caches some information) or when
away-notify exists.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>