ui: add highest channel membership prefix to nicklist
Adds the highest channel membership prefix flag (e.g. @ or +) next to
each user's nick in the nicklist.

Since prefixes are specific to user+channel combination, we add a Member
struct that contains a pointer to a User and a prefix character, and
each Channel then contains a list of Members. For now, the prefix is
only a single character (u8) which contains the highest prefix for that
user (as determined from the WHOX or NAMES reply). If comlink ever
enables the multi-prefix extension we may want to change this to be a
string, but since we only display a single character I think this is ok
for now.