client: fix memory leak from should_close race
When exiting the client, the read loop could create a memory leak by
never closing. If a message is received and partially processed, but not
enterred into the event loop it will be a memory leak. Change the while
predicate to be !should_close, which ensures that the read loop will
properly exit. The exit flow for a client is:

1. Set should_close = true
2. Write PING to the client to trigger a PONG response
3. Client responds with PONG
4. The read loop reads the PONG and exits

Prior to this commit, the read loop would not properly exit

Fixes: https://github.com/rockorager/comlink/issues/3