ui: add highlight to unread messages
All unread messages are highlighted with a gray background. This is
currently the same color used on mouse hover, simply because it was easy
to copy.

A channel is marked read when:

* You switch channel (either via mouse or keyboard); or
* you send a message.

This means that new messages _from others in the channel_ will remain
unread until you either:

* send a message; or
* change the current channel.

Future improvements:

* Add a keybinding to mark the channel read (easy-ish)
* Mark a message as read as soon as it's received if the terminal window
  (and thus comlink itself) is focused (harder).
* Mark a channel read on window focus (easy-ish).
    - This might actually not be nice? I wouldn't want a channel to be
      marked read as soon as I focus the window in case there's a long
      scrollback I want to catch up on.